Add files via upload
This script predicts temperature values in unknown wells using the Gradient Boosting Regressor machine learning technique. Spatial weights to the closest wells are incorporated into the prediction. The predicted temperature gradients are cleaned of noise by fitting a regression line. The results can be visualized on a plot and are also available in the final DataFrame.